cmd/sbig-info: refactor info subcommand
Display each camera info structure in separate functions,
treating CE_BAD_PARAMETER as a non-fatal error meaning
"request inappropriate for this camera model" in cases
where the info structure may not be supported, rather
than hardwiring that based on model/chip.

I wasn't sure what to do with the STX/STXL bit in info6,
so I just grab info6 in the info0 function if the camera
shows as STX, and display it as STXL if the bit is set.