cmd/sbig-snap: --device should override config file
Problem: "sbig --device XYZ snap" doesn't override the device
in the config file, which is different than the other tools.

The sbig command driver first reads the device setting from
the config file, then allows it to be overridden with --device,
before setting SBIG_DEVICE and executing the subcommand.  In
subcommands, the SBIG_DEVICE setting should be used directly.

Fix sbig-snap to ignore the device setting in the config file
when it pulls out other setup info.

Fixes #26