libsbig/camera: ST-5C/255: use cfw as shutter
Problem: on model ST-5C/255, shutter directives are
ingored and sbig-snap in "autodark" mode subtracts LF
from LF, canceling the signal.

Add a special case for 5C/255 in sbig_ccd_start_exposure()
and sbig_ccd_end_exposure() that compensates for the lack
of shutter with the internal filter wheel.  The caller just
specifies the shutter directives as usual.

Fixes #29