libsbig/handle: dlopen libsbigudrv.so
Problem: INDI-packagerd libsbigudrv.so was not
found at runtime on i386 architecture.

The strategy of adding -lsbigudrv to LIBS and then
dlopening NULL and using dlsym to find library paths
was in retrospect fragile.  It worked on arm7l but
does not on i386.

Skip even searching for libsbigudrv.so at configure
time.  If the user doesn't explicitly use sbig -S path,
then dlopen the relative path "libsbigudrv.so" and
let the system search algorithm described in dlopen(3)
find it.