libsbig/handle: change default dlopen path (again)
Problem: INDI renamed the sbigudrv library in their 3rd party
debs from libsbigudrv.so.2.1.1 to libsbig.so.2.1.3 and now we
can't find it.

The INDI deb has fixed the symlink issue worked around in
pr #40, so drop the explicit version from the library name.
The basename has also changed so update that.

Note that the underlying binary release from SBIG/DL is still at
version 4.84 (when interrogated) so despite the shared library
version number change, the ABI should be unchanged.

On a system that still has the old packages installed,
the path can overridden in ~/.sbig/config.ini, e.g.

[system]
sbigudrv = libsbigudrv.so.2.1.1